Simplify sending backups by email.
By using a different email provider that allows more
emails per month I could remove the complicated logic
to distribute the emails on the days of the month.

Now everybody gets the backup on the first day of the
month. The hour depends on the hour a user registed on Yith