Redirect insecure requests to HTTPS scheme
Platforms like Heroku are already offering a TLS endpoint in
addition to the plain HTTP scheme. But since both schemes work
independently from each other, any redirection must be done in
the application server.

This patch solves that adding a new middleware in order to inspect
for every connection some of the HTTP headers sent from the load
balancer (in the case of Heroku).

Based on this post:
http://brandon-bradley.com/blog/2014/11/heroku-piggyback-ssl-for-pyramid-1-5/